Framerates are floats

+++ DrainLib/Engines/BaseEngineAccessor.cs
@@ -103,7 +103,7 @@ namespace DrainLib.Engines {
 	public class VideoState {
 		public uint CurrentFrame;
 		public uint FrameCount;
-		public int FrameRate;
+		public float FrameRate;
 		public string FileName;
 	}
 }
